src/server/requirements.txt
Flask==1.1.2
pandas==1.3.2
numpy==1.18.1
sqlalchemy==1.4.15 
psycopg2-binary==2.8.4
xlrd==1.2.0  # currently used for xlsx, but we should consider adjusting code to openpyxl for xlsx
openpyxl
requests
pytest
flask-jwt-extended==4.0.2
alembic
flask-cors
dropbox
jellyfish
cryptography